Moved share script into it's own file
