+++ lib/modules/dosomething/dosomething_reportback/dosomething_reportback.module
@@ -461,8 +461,11 @@ function dosomething_reportback_view_entity($entity) {
     $subtitle = ($is_owner) ? $copy_vars['owners_rb_subtitle'] . " " . $incentive : $node->call_to_action;
     $link = '/' . drupal_get_path_alias('node/' . $node->nid);
 
-    if (strlen($entity->why_participated) > 250) {
-      $why_participated_short = dosomething_helpers_text_truncate($entity->why_participated, 250, '...', FALSE, FALSE);
+    $character_limit = 250;
+    $participated_copy = $entity->why_participated;
+
+    if (strlen($participated_copy) > $character_limit) {
+      $why_participated_short = dosomething_helpers_text_truncate($participated_copy, $character_limit, '...', FALSE, FALSE);
     }
 
     $reportback_url = url(current_path(), array('absolute' => TRUE, 'query' => array('fid' => $fid)));
@@ -505,7 +508,7 @@ function dosomething_reportback_view_entity($entity) {
     );
     cache_set('ds_permalink_' . $entity->rbid . '_' . $fid . '_' . $is_owner, $permalink, 'cache_dosomething_reportback');
     return $permalink;
-  // }
+  }
 }
 /**
  * Custom submit handler for signup form.
