+++ lib/themes/dosomething/paraneue_dosomething/js/app.js
@@ -13,6 +13,7 @@ define("app", function (require) {
   var Revealer = require("revealer/Revealer");
   var Swapper  = require("swapper/Swapper");
   var Reportback = require("reportback/Reportback");
+  var Share    = require("utils/Share.js");
 
   // Initialize modules on load
   require("neue");
@@ -96,34 +97,37 @@ define("app", function (require) {
       Reportback.init($reportback);
     }
 
+    if ($body.find("a.js-share-link").length) {
+      Share.popup();
+    }
 
-    // Open share links in a popup dialog.
-    var windowOptions = 'scrollbars=yes,resizable=yes,toolbar=no,location=yes';
-    var width         = 550;
-    var height        = 420;
-    var winHeight     = screen.height;
-    var winWidth      = screen.width;
+    // // Open share links in a popup dialog.
+    // var windowOptions = 'scrollbars=yes,resizable=yes,toolbar=no,location=yes';
+    // var width         = 550;
+    // var height        = 420;
+    // var winHeight     = screen.height;
+    // var winWidth      = screen.width;
 
-    $('body').on('click', 'a.js-share-link', function(event) {
-      event.preventDefault();
-      var leftPos = Math.round((winWidth / 2) - (width / 2));
-      var topPos = 0;
+    // $('body').on('click', 'a.js-share-link', function(event) {
+    //   event.preventDefault();
+    //   var leftPos = Math.round((winWidth / 2) - (width / 2));
+    //   var topPos = 0;
 
-      if (winHeight > height) {
-        topPos = Math.round((winHeight / 2) - (height / 2));
-      }
+    //   if (winHeight > height) {
+    //     topPos = Math.round((winHeight / 2) - (height / 2));
+    //   }
 
-      window.open(this.href, '_blank', windowOptions + ',width=' + width +
-       ',height=' + height + ',left=' + leftPos + ',top=' + topPos);
-    });
+    //   window.open(this.href, '_blank', windowOptions + ',width=' + width +
+    //    ',height=' + height + ',left=' + leftPos + ',top=' + topPos);
+    // });
 
-    $participate = $body.find(".js-permalink-show-more");
+    $showMoreText = $body.find(".js-permalink-show-more");
 
-    if ($participate.length) {
-      $participate.click(function(event) {
+    if ($showMoreText.length) {
+      $showMoreText.click(function(event) {
         event.preventDefault();
-        $('.participate').toggleClass("is-shown");
-        ($participate.html() === "Show More") ? $participate.html("Show Less") : $participate.html("Show More");
+        $(".participate").toggleClass("is-shown");
+        ($showMoreText.html() === "Show More") ? $showMoreText.html("Show Less") : $showMoreText.html("Show More");
       });
     }
